chore: move walletModal to ui-components, update prop types
