+++ apps/bridge-dapp/src/components/Header/index.ts
@@ -1,2 +1,1 @@
 export * from './Header';
-export * from './WalletModal';
