+++ apps/bridge-dapp/src/containers/TransferConfirmContainer/types.ts
@@ -121,4 +121,9 @@ export interface TransferConfirmContainerProps
    * The refund token
    */
   refundToken?: string;
+
+  /**
+   * The active chain
+   */
+  activeChain: Chain;
 }
