+++ apps/bridge-dapp/src/containers/WalletModalContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './WalletModalContainer';
