+++ apps/tangle-dapp/containers/index.ts
@@ -3,3 +3,4 @@ export * from './KeyMetricsTableContainer';
 export { Layout } from './Layout';
 export * from './ValidatorTablesContainer';
 export * from './WalletAndChainContainer';
+export * from './WalletModalContainer';
