+++ libs/webb-ui-components/src/components/ConnectWalletMobileButton/types.ts
@@ -1,15 +1,12 @@
 import { ButtonProps } from '../buttons/types';
 import { IWebbComponentBase } from '../../types';
 
-export type MobileButtonProps = {
-  title?: string;
-  extraActionButtons?: Array<ButtonProps>;
-};
-
 export interface ConnectWalletMobileButtonProps
   extends IWebbComponentBase,
     Pick<
       ButtonProps,
       'variant' | 'isFullWidth' | 'leftIcon' | 'rightIcon' | 'size' | 'variant'
-    >,
-    MobileButtonProps {}
+    > {
+  title?: string;
+  extraActionButtons?: Array<ButtonProps>;
+}
