+++ libs/webb-ui-components/src/components/WalletModal/index.ts
