+++ libs/webb-ui-components/src/components/WalletModal/types.ts
@@ -1,0 +1,64 @@
+import { ApiConfig, WalletConfig } from '@webb-tools/dapp-config';
+import { notificationApi } from '../Notification';
+import { PropsOf } from '../../types';
+import { WalletId, WebbError } from '@webb-tools/dapp-types';
+
+export interface WalletModalProps extends PropsOf<'div'> {
+  /**
+   * The api to use for notifications
+   */
+  notificationApi: typeof notificationApi;
+
+  /**
+   * The api config to use for the wallet modal
+   */
+  apiConfig: ApiConfig;
+
+  /**
+   * The wallet id that is currently connecting
+   */
+  connectingWalletId?: WalletId;
+
+  /**
+   * The wallet id that has failed to connect
+   */
+  failedWalletId?: WalletId;
+
+  /**
+   * Is the modal open
+   */
+  isModalOpen: boolean;
+
+  /**
+   * Reset the state of the modal
+   */
+  resetState: () => void;
+
+  /**
+   * The selected wallet
+   */
+  selectedWallet?: WalletConfig;
+
+  /**
+   * Connect the wallet
+   */
+  connectWallet: (wallet: WalletConfig) => void | Promise<void>;
+
+  /**
+   * Toggle the modal
+   */
+  toggleModal: (
+    isOpen?: boolean | undefined,
+    typedChainId?: number | undefined
+  ) => void;
+
+  /**
+   * The error that has occurred when connecting
+   */
+  connectError?: WebbError;
+
+  /**
+   * The wallets to display in the modal
+   */
+  supportedWallets: WalletConfig[];
+}
