+++ libs/webb-ui-components/src/components/index.ts
@@ -84,6 +84,7 @@ export * from './TxProgressor';
 export { default as TxProgressor } from './TxProgressor';
 export { default as TxConfirmationRing } from './TxConfirmationRing';
 export * from './WalletConnectionCard';
+export * from './WalletModal';
 export * from './WebsiteCommunity';
 export * from './WebsiteFooter';
 export * from './WebsiteNewsLetterForm';
